// src/enums/feature-type.ts
export enum FeatureType {
	Text = 'Текст',
	Ability = 'Здібність',
	AbilityCost = 'Вартість здібності',
	AbilityDamage = 'Шкода здібності',
	AbilityDistance = 'Дальність здібності',
	AddOn = 'Додаток',
	AncestryChoice = 'Вибір походження',
	AncestryFeatureChoice = 'Вибір особливості походження',
	Bonus = 'Бонус',
	CharacteristicBonus = 'Бонус характеристики',
	Choice = 'Вибір',
	ClassAbility = 'Класова здібність',
	Companion = 'Компаньйон',
	ConditionImmunity = 'Імунітет до станів',
	DamageModifier = 'Модифікатор шкоди',
	Domain = 'Домен',
	DomainFeature = 'Особливість домену',
	Fixture = 'Фікстура',
	Follower = 'Прибічник',
	HeroicResource = 'Героїчний ресурс',
	HeroicResourceGain = 'Отримання героїчного ресурсу',
	ItemChoice = 'Вибір предмета',
	Kit = 'Набір',
	Language = 'Мова',
	LanguageChoice = 'Вибір мови',
	Malice = 'Злоба',
	MaliceAbility = 'Здібність злоби',
	MovementMode = 'Режим руху',
	Multiple = 'Набір особливостей',
	Package = 'Пакет',
	PackageContent = 'Вміст пакета',
	Perk = 'Перевага',
	Proficiency = 'Володіння',
	Retainer = 'Ретейнер',
	SaveThreshold = 'Поріг порятунку',
	Size = 'Розмір',
	SkillChoice = 'Вибір навички',
	Speed = 'Швидкість',
	Summon = 'Виклик',
	SummonChoice = 'Вибір виклику',
	TaggedFeature = 'Позначена особливість',
	TaggedFeatureChoice = 'Вибір позначеної особливості',
	TitleChoice = 'Вибір титулу'
}